# springboot-httpbin-reactive/src/main/java/com/dyrnq/httpbin/api/AnythingApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.9.0-SNAPSHOT).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.dyrnq.httpbin.api;

import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.multipart.MultipartFile;
import org.springframework.web.server.ServerWebExchange;
import reactor.core.publisher.Flux;
import reactor.core.publisher.Mono;
import org.springframework.http.codec.multipart.Part;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.9.0-SNAPSHOT")
@Validated
@Tag(name = "Anything", description = "Returns anything that is passed to request")
public interface AnythingApi {

    /**
     * DELETE /anything/{anything} : Returns anything passed in request data.
     *
     * @param anything  (required)
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingAnythingDelete",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/anything/{anything}"
    )
    
    default Mono<ResponseEntity<Void>> anythingAnythingDelete(
        @Parameter(name = "anything", description = "", required = true, in = ParameterIn.PATH) @PathVariable("anything") String anything,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * GET /anything/{anything} : Returns anything passed in request data.
     *
     * @param anything  (required)
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingAnythingGet",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/anything/{anything}"
    )
    
    default Mono<ResponseEntity<Void>> anythingAnythingGet(
        @Parameter(name = "anything", description = "", required = true, in = ParameterIn.PATH) @PathVariable("anything") String anything,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * PATCH /anything/{anything} : Returns anything passed in request data.
     *
     * @param anything  (required)
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingAnythingPatch",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/anything/{anything}"
    )
    
    default Mono<ResponseEntity<Void>> anythingAnythingPatch(
        @Parameter(name = "anything", description = "", required = true, in = ParameterIn.PATH) @PathVariable("anything") String anything,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * POST /anything/{anything} : Returns anything passed in request data.
     *
     * @param anything  (required)
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingAnythingPost",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/anything/{anything}"
    )
    
    default Mono<ResponseEntity<Void>> anythingAnythingPost(
        @Parameter(name = "anything", description = "", required = true, in = ParameterIn.PATH) @PathVariable("anything") String anything,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * PUT /anything/{anything} : Returns anything passed in request data.
     *
     * @param anything  (required)
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingAnythingPut",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/anything/{anything}"
    )
    
    default Mono<ResponseEntity<Void>> anythingAnythingPut(
        @Parameter(name = "anything", description = "", required = true, in = ParameterIn.PATH) @PathVariable("anything") String anything,
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * DELETE /anything : Returns anything passed in request data.
     *
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingDelete",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = "/anything"
    )
    
    default Mono<ResponseEntity<Void>> anythingDelete(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * GET /anything : Returns anything passed in request data.
     *
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingGet",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/anything"
    )
    
    default Mono<ResponseEntity<Void>> anythingGet(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * PATCH /anything : Returns anything passed in request data.
     *
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingPatch",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.PATCH,
        value = "/anything"
    )
    
    default Mono<ResponseEntity<Void>> anythingPatch(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * POST /anything : Returns anything passed in request data.
     *
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingPost",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/anything"
    )
    
    default Mono<ResponseEntity<Void>> anythingPost(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }


    /**
     * PUT /anything : Returns anything passed in request data.
     *
     * @return Anything passed in request (status code 200)
     */
    @Operation(
        operationId = "anythingPut",
        summary = "Returns anything passed in request data.",
        tags = { "Anything" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Anything passed in request")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = "/anything"
    )
    
    default Mono<ResponseEntity<Void>> anythingPut(
        @Parameter(hidden = true) final ServerWebExchange exchange
    ) throws Exception {
        Mono<Void> result = Mono.empty();
        exchange.getResponse().setStatusCode(HttpStatus.NOT_IMPLEMENTED);
        return result.then(Mono.empty());

    }

}