# springboot-httpbin/src/main/resources/openapi.yaml
openapi: 3.0.1
info:
  contact:
    email: dyrnq@outlook.com
    url: https://github.com/dyrnq
  description: springboot-httpbin
  title: springboot-httpbin
  version: 0.0.1
servers:
- url: https://httpbin.org/
...
paths:
...
  /anything/{anything}:
    delete:
      parameters:
      - explode: false
        in: path
        name: anything
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Anything passed in request
      summary: Returns anything passed in request data.
      tags:
      - Anything
      x-accepts:
      - application/json
      x-tags:
      - tag: Anything
    get:
      parameters:
      - explode: false
        in: path
        name: anything
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Anything passed in request
      summary: Returns anything passed in request data.
      tags:
      - Anything
      x-accepts:
      - application/json
      x-tags:
      - tag: Anything
    patch:
      parameters:
      - explode: false
        in: path
        name: anything
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Anything passed in request
      summary: Returns anything passed in request data.
      tags:
      - Anything
      x-accepts:
      - application/json
      x-tags:
      - tag: Anything
    post:
      parameters:
      - explode: false
        in: path
        name: anything
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Anything passed in request
      summary: Returns anything passed in request data.
      tags:
      - Anything
      x-accepts:
      - application/json
      x-tags:
      - tag: Anything
    put:
      parameters:
      - explode: false
        in: path
        name: anything
        required: true
        schema:
          type: string
        style: simple
      responses:
        "200":
          content: {}
          description: Anything passed in request
      summary: Returns anything passed in request data.
      tags:
      - Anything
      x-accepts:
      - application/json
      x-tags:
      - tag: Anything
...